chore: move workflow stuff around
